requirements.txt
boto3==1.34.116
pandas==2.2.2
streamlit==1.35.0
mysql-connector-python==8.4.0
SQLAlchemy==1.4
google-generativeai==0.2.0
openai==0.27.0